Add a test case for all WFDB FLAC signal formats.
The record "flacformats" contains one signal in each of the three WFDB
FLAC signal formats.  As with the "binformats" record, sample j of
signal i is equal to:

   (i + 16843019 * j) % ((1 << adcres) - 1) + 1 - (1 << (adcres - 1)))

Use this record to test that it is possible to read all of the formats
correctly, including when we skip one or two samples from the start
and/or end of the record.