Add a test case for reading FLAC files over HTTP.
We want to be sure that the soundfile library is compatible with
wfdb.io._url.NetFile, since wfdb.io._signal._rd_compressed_file relies
on this.

This isn't a complete test of WFDB/FLAC functionality, in part because
the sample files are very small so we can't really exercise the
seeking functionality.  However, this should be sufficient to verify
that soundfile doesn't make silly assumptions (like assuming the input
file corresponds to an OS file descriptor.)